// admin/src/identity-provider/interface/identity-provider.interface.ts
export interface IIdentityProvider {
  id?: string;
  uid: string;
  name: string;
  issuer: string;
  messageToDisplayWhenInactive: string;
  redirectionTargetWhenInactive: string;
  active: boolean;
  display: boolean;
  isBeta: boolean;
  title: string;
  image: string;
  alt: string;
  imageFocus: string;
  eidas: number;
  allowedAcr?: string[];
  emails?: string[];
  supportEmail?: string;
  statusUrl: string;
  authorizationUrl?: string;
  tokenUrl?: string;
  userInfoUrl?: string;
  logoutUrl: string;
  clientId: string;
  client_secret: string;
  order: number;
  jwksUrl?: string;
  jwtAlgorithm: string[];
  discoveryUrl?: string;
  discovery: boolean;
  trustedIdentity: boolean;
  specificText: string;
  createdAt: Date;
  updatedAt: Date;
  updatedBy: string;
  blacklistByIdentityProviderActivated: boolean;
  whitelistByServiceProviderActivated: boolean;
  userinfo_encrypted_response_enc?: string;
  userinfo_encrypted_response_alg?: string;
  userinfo_signed_response_alg?: 'RS256' | 'ES256' | 'HS256' | '';
  id_token_signed_response_alg?: 'RS256' | 'ES256' | 'HS256';
  id_token_encrypted_response_alg?: string;
  id_token_encrypted_response_enc?: string;
  token_endpoint_auth_method?: string;
  revocation_endpoint_auth_method?: string;
  amr?: string[];
  modalActive?: boolean;
  modalTitle?: string;
  modalBody?: string;
  modalContinueText?: string;
  modalMoreInfoLabel?: string;
  modalMoreInfoUrl?: string;
}